requirements.txt
pyTelegramBotAPI
python-dotenv
requests==2.26.0
Pillow==8.4.0
googletrans==4.0.0rc1
